requirements.txt
flask
flask-cors
requests
speedtest-cli
